Added tests to test_single_project.ts.
