+++ tests/test_manage_projects/test_manage_projects.ts
@@ -1,0 +1,133 @@
+import ManageProjects from '../../src/components/manage_projects/manage_projects.vue';
+import ValidatedInput from '../../src/components/validated_input.vue';
+import { config, mount, RouterLinkStub, Wrapper } from '@vue/test-utils';
+import { Course, Project, Semester , UltimateSubmissionPolicy } from 'ag-client-typescript';
+import { AxiosError, AxiosResponse } from 'axios';
+import Vue from 'vue';
+
+import {
+    patch_async_class_method,
+    patch_async_static_method,
+    patch_component_method
+} from '../mocking';
+
+beforeAll(() => {
+    config.logModifiedComponents = false;
+});
+
+describe('ManageProjects.vue', () => {
+    let wrapper: Wrapper<ManageProjects>;
+    let course: Course;
+    let original_match_media: (query: string) => MediaQueryList;
+    let new_project: Project;
+    let project_1: Project;
+    let project_2: Project;
+    let projects: Project[];
+
+    const $route = {
+        path: '/web/course_admin/:courseId',
+        params: { courseId: '2' }
+    };
+
+
+    beforeEach(() => {
+        course = new Course({
+            pk: 2, name: 'EECS 280', semester: Semester.winter, year: 2019, subtitle: '',
+            num_late_days: 0, allowed_guest_domain: '', last_modified: ''
+        });
+        original_match_media = window.matchMedia;
+        Object.defineProperty(window, "matchMedia", {
+            value: jest.fn(() => {
+                return {matches: true};
+            })
+        });
+
+        project_1 = new Project({
+            pk: 10,
+            name: "Project 1 - Statistics",
+            last_modified: "today",
+            course: 2,
+            visible_to_students: true,
+            closing_time: null,
+            soft_closing_time: null,
+            disallow_student_submissions: true,
+            disallow_group_registration: true,
+            guests_can_submit: true,
+            min_group_size: 1,
+            max_group_size: 1,
+            submission_limit_per_day: null,
+            allow_submissions_past_limit: true,
+            groups_combine_daily_submissions: false,
+            submission_limit_reset_time: "",
+            submission_limit_reset_timezone: "",
+            num_bonus_submissions: 1,
+            total_submission_limit: null,
+            allow_late_days: true,
+            ultimate_submission_policy: UltimateSubmissionPolicy.best,
+            hide_ultimate_submission_fdbk: false
+        });
+
+        project_2 = new Project({
+            pk: 4,
+            name: "Project 2 - Images",
+            last_modified: "today",
+            course: 2,
+            visible_to_students: true,
+            closing_time: null,
+            soft_closing_time: null,
+            disallow_student_submissions: true,
+            disallow_group_registration: true,
+            guests_can_submit: true,
+            min_group_size: 1,
+            max_group_size: 1,
+            submission_limit_per_day: null,
+            allow_submissions_past_limit: true,
+            groups_combine_daily_submissions: false,
+            submission_limit_reset_time: "",
+            submission_limit_reset_timezone: "",
+            num_bonus_submissions: 1,
+            total_submission_limit: null,
+            allow_late_days: true,
+            ultimate_submission_policy: UltimateSubmissionPolicy.best,
+            hide_ultimate_submission_fdbk: false
+        });
+
+        projects = [project_1, project_2];
+
+        new_project = new Project({
+            pk: 12,
+            name: "Lab 09 - Functors",
+            last_modified: "today",
+            course: 2,
+            visible_to_students: true,
+            closing_time: null,
+            soft_closing_time: null,
+            disallow_student_submissions: true,
+            disallow_group_registration: true,
+            guests_can_submit: true,
+            min_group_size: 1,
+            max_group_size: 1,
+            submission_limit_per_day: null,
+            allow_submissions_past_limit: true,
+            groups_combine_daily_submissions: false,
+            submission_limit_reset_time: "",
+            submission_limit_reset_timezone: "",
+            num_bonus_submissions: 1,
+            total_submission_limit: null,
+            allow_late_days: true,
+            ultimate_submission_policy: UltimateSubmissionPolicy.best_with_normal_fdbk,
+            hide_ultimate_submission_fdbk: false
+        });
+    });
+
+    afterEach(() => {
+        Object.defineProperty(window, "matchMedia", {
+            value: original_match_media
+        });
+
+        if (wrapper.exists()) {
+            console.log("wrapper exists");
+            wrapper.destroy();
+        }
+    });
+});
